+++ NEWS.md
@@ -1,8 +1,19 @@
-# ggnormviolin 0.1.2
+# ggnormalviolin 0.2.1
 
-* Fixed bug preventing ggnormalviolin from working without loading ggplot2 first
+# ggnormalviolin 0.2.0 _2025-03-24_
 
+- Changed default aesthetic `size` to `linewidth`
+- Updated documentation
 
-# ggnormviolin 0.1.1
+# ggnormalviolin 0.1.3
 
-* Initial release
+-   Added test coverage
+-   Revised code according to recommendations in goodpractice::gp()
+
+# ggnormalviolin 0.1.2
+
+-   Fixed bug preventing ggnormalviolin from working without loading ggplot2 first
+
+# ggnormalviolin 0.1.1
+
+-   Initial release